+++ src/Controller/UserController.php
@@ -4,7 +4,9 @@ namespace App\Controller;
 
 use App\Entity\User;
 use App\Form\User1Type;
+use App\Form\UserType;
 use App\Repository\UserRepository;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,10 +20,20 @@ class UserController extends AbstractController
     /**
      * @Route("/", name="app_user_index", methods={"GET"})
      */
-    public function index(UserRepository $userRepository): Response
+    public function index(UserRepository $userRepository,Request $request,PaginatorInterface $paginator): Response
     {
+        if (!($this->getUser())) {
+            return $this->redirectToRoute('app_home');
+        }
+        if(!(in_array("ROLE_ADMIN",$this->getUser()->getRoles()))) {
+            return $this->redirectToRoute('app_home');
+        }
+        $search = $request->query->get("search");
+        $users = $userRepository->findAllWithSearch($search);
+        //$users= $userRepository->findAll();
+
         return $this->render('user/index.html.twig', [
-            'users' => $userRepository->findAll(),
+            'users' => $users,
         ]);
     }
 
@@ -30,6 +42,13 @@ class UserController extends AbstractController
      */
     public function new(Request $request, UserRepository $userRepository): Response
     {
+        if (!($this->getUser())) {
+            return $this->redirectToRoute('app_home');
+        }
+
+        if(!(in_array("ROLE_ADMIN",$this->getUser()->getRoles()))) {
+            return $this->redirectToRoute('app_home');
+        }
         $user = new User();
         $form = $this->createForm(User1Type::class, $user);
         $form->handleRequest($request);
@@ -50,6 +69,13 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
+
+        if(!(in_array("ROLE_ADMIN",$this->getUser()->getRoles()))) {
+            return $this->redirectToRoute('app_home');
+        }
+        if (!($this->getUser())) {
+            return $this->redirectToRoute('app_home');
+        }
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -60,7 +86,14 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
-        $form = $this->createForm(User1Type::class, $user);
+        if (!($this->getUser())) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        if(!(in_array("ROLE_ADMIN",$this->getUser()->getRoles()))) {
+            return $this->redirectToRoute('app_home');
+        }
+        $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -79,10 +112,23 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        if (!($this->getUser())) {
+            return $this->redirectToRoute('app_login');
+        }
+
+        if(!(in_array("ROLE_ADMIN",$this->getUser()->getRoles()))) {
+            return $this->redirectToRoute('app_home');
+        }
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user);
         }
 
         return $this->redirectToRoute('app_user_index', [], Response::HTTP_SEE_OTHER);
     }
+
+
+
+
+
+
 }
